feat: changed interceptor to an specific class
